fix: Separated service namespaces from module entry points
