+++ etc/firebase-admin.auth.api.md
@@ -29,115 +29,6 @@ export class Auth extends BaseAuth {
     }
 
 // @public
-export function auth(app?: App): auth.Auth;
-
-// @public (undocumented)
-export namespace auth {
-    // (undocumented)
-    export type ActionCodeSettings = ActionCodeSettings;
-    // (undocumented)
-    export type Auth = Auth;
-    // (undocumented)
-    export type AuthFactorType = AuthFactorType;
-    // (undocumented)
-    export type AuthProviderConfig = AuthProviderConfig;
-    // (undocumented)
-    export type AuthProviderConfigFilter = AuthProviderConfigFilter;
-    // (undocumented)
-    export type BaseAuth = BaseAuth;
-    // (undocumented)
-    export type CreateMultiFactorInfoRequest = CreateMultiFactorInfoRequest;
-    // (undocumented)
-    export type CreatePhoneMultiFactorInfoRequest = CreatePhoneMultiFactorInfoRequest;
-    // (undocumented)
-    export type CreateRequest = CreateRequest;
-    // (undocumented)
-    export type CreateTenantRequest = CreateTenantRequest;
-    // (undocumented)
-    export type DecodedIdToken = DecodedIdToken;
-    // (undocumented)
-    export type DeleteUsersResult = DeleteUsersResult;
-    // (undocumented)
-    export type EmailIdentifier = EmailIdentifier;
-    // (undocumented)
-    export type EmailSignInProviderConfig = EmailSignInProviderConfig;
-    // (undocumented)
-    export type GetUsersResult = GetUsersResult;
-    // (undocumented)
-    export type HashAlgorithmType = HashAlgorithmType;
-    // (undocumented)
-    export type ListProviderConfigResults = ListProviderConfigResults;
-    // (undocumented)
-    export type ListTenantsResult = ListTenantsResult;
-    // (undocumented)
-    export type ListUsersResult = ListUsersResult;
-    // (undocumented)
-    export type MultiFactorConfig = MultiFactorConfig;
-    // (undocumented)
-    export type MultiFactorConfigState = MultiFactorConfigState;
-    // (undocumented)
-    export type MultiFactorCreateSettings = MultiFactorCreateSettings;
-    // (undocumented)
-    export type MultiFactorInfo = MultiFactorInfo;
-    // (undocumented)
-    export type MultiFactorSettings = MultiFactorSettings;
-    // (undocumented)
-    export type MultiFactorUpdateSettings = MultiFactorUpdateSettings;
-    // (undocumented)
-    export type OIDCAuthProviderConfig = OIDCAuthProviderConfig;
-    // (undocumented)
-    export type OIDCUpdateAuthProviderRequest = OIDCUpdateAuthProviderRequest;
-    // (undocumented)
-    export type PhoneIdentifier = PhoneIdentifier;
-    // (undocumented)
-    export type PhoneMultiFactorInfo = PhoneMultiFactorInfo;
-    // (undocumented)
-    export type ProviderIdentifier = ProviderIdentifier;
-    // (undocumented)
-    export type SAMLAuthProviderConfig = SAMLAuthProviderConfig;
-    // (undocumented)
-    export type SAMLUpdateAuthProviderRequest = SAMLUpdateAuthProviderRequest;
-    // (undocumented)
-    export type SessionCookieOptions = SessionCookieOptions;
-    // (undocumented)
-    export type Tenant = Tenant;
-    // (undocumented)
-    export type TenantAwareAuth = TenantAwareAuth;
-    // (undocumented)
-    export type TenantManager = TenantManager;
-    // (undocumented)
-    export type UidIdentifier = UidIdentifier;
-    // (undocumented)
-    export type UpdateAuthProviderRequest = UpdateAuthProviderRequest;
-    // (undocumented)
-    export type UpdateMultiFactorInfoRequest = UpdateMultiFactorInfoRequest;
-    // (undocumented)
-    export type UpdatePhoneMultiFactorInfoRequest = UpdatePhoneMultiFactorInfoRequest;
-    // (undocumented)
-    export type UpdateRequest = UpdateRequest;
-    // (undocumented)
-    export type UpdateTenantRequest = UpdateTenantRequest;
-    // (undocumented)
-    export type UserIdentifier = UserIdentifier;
-    // (undocumented)
-    export type UserImportOptions = UserImportOptions;
-    // (undocumented)
-    export type UserImportRecord = UserImportRecord;
-    // (undocumented)
-    export type UserImportResult = UserImportResult;
-    // (undocumented)
-    export type UserInfo = UserInfo;
-    // (undocumented)
-    export type UserMetadata = UserMetadata;
-    // (undocumented)
-    export type UserMetadataRequest = UserMetadataRequest;
-    // (undocumented)
-    export type UserProviderRequest = UserProviderRequest;
-    // (undocumented)
-    export type UserRecord = UserRecord;
-}
-
-// @public
 export type AuthFactorType = 'phone';
 
 // @public
