+++ etc/firebase-admin.messaging.api.md
@@ -194,79 +194,6 @@ export class Messaging {
     }
 
 // @public
-export function messaging(app?: App): messaging.Messaging;
-
-// @public (undocumented)
-export namespace messaging {
-    // (undocumented)
-    export type AndroidConfig = AndroidConfig;
-    // (undocumented)
-    export type AndroidFcmOptions = AndroidFcmOptions;
-    // (undocumented)
-    export type AndroidNotification = AndroidNotification;
-    // (undocumented)
-    export type ApnsConfig = ApnsConfig;
-    // (undocumented)
-    export type ApnsFcmOptions = ApnsFcmOptions;
-    // (undocumented)
-    export type ApnsPayload = ApnsPayload;
-    // (undocumented)
-    export type Aps = Aps;
-    // (undocumented)
-    export type ApsAlert = ApsAlert;
-    // (undocumented)
-    export type BatchResponse = BatchResponse;
-    // (undocumented)
-    export type ConditionMessage = ConditionMessage;
-    // (undocumented)
-    export type CriticalSound = CriticalSound;
-    // (undocumented)
-    export type DataMessagePayload = DataMessagePayload;
-    // (undocumented)
-    export type FcmOptions = FcmOptions;
-    // (undocumented)
-    export type LightSettings = LightSettings;
-    // (undocumented)
-    export type Message = Message;
-    // (undocumented)
-    export type Messaging = Messaging;
-    // (undocumented)
-    export type MessagingConditionResponse = MessagingConditionResponse;
-    // (undocumented)
-    export type MessagingDeviceGroupResponse = MessagingDeviceGroupResponse;
-    // (undocumented)
-    export type MessagingDeviceResult = MessagingDeviceResult;
-    // (undocumented)
-    export type MessagingDevicesResponse = MessagingDevicesResponse;
-    // (undocumented)
-    export type MessagingOptions = MessagingOptions;
-    // (undocumented)
-    export type MessagingPayload = MessagingPayload;
-    // (undocumented)
-    export type MessagingTopicManagementResponse = MessagingTopicManagementResponse;
-    // (undocumented)
-    export type MessagingTopicResponse = MessagingTopicResponse;
-    // (undocumented)
-    export type MulticastMessage = MulticastMessage;
-    // (undocumented)
-    export type Notification = Notification;
-    // (undocumented)
-    export type NotificationMessagePayload = NotificationMessagePayload;
-    // (undocumented)
-    export type SendResponse = SendResponse;
-    // (undocumented)
-    export type TokenMessage = TokenMessage;
-    // (undocumented)
-    export type TopicMessage = TopicMessage;
-    // (undocumented)
-    export type WebpushConfig = WebpushConfig;
-    // (undocumented)
-    export type WebpushFcmOptions = WebpushFcmOptions;
-    // (undocumented)
-    export type WebpushNotification = WebpushNotification;
-}
-
-// @public
 export interface MessagingConditionResponse {
     messageId: number;
 }
