+++ src/auth/index.ts
@@ -14,6 +14,40 @@
  * limitations under the License.
  */
 
+import { App, getApp } from '../app/index';
+import { FirebaseApp } from '../app/firebase-app';
+import { Auth } from './auth';
+
+/**
+ * Gets the {@link auth.Auth `Auth`} service for the default app or a
+ * given app.
+ *
+ * `getAuth()` can be called with no arguments to access the default app's
+ * {@link auth.Auth `Auth`} service or as `getAuth(app)` to access the
+ * {@link auth.Auth `Auth`} service associated with a specific app.
+ *
+ * @example
+ * ```javascript
+ * // Get the Auth service for the default app
+ * const defaultAuth = getAuth();
+ * ```
+ *
+ * @example
+ * ```javascript
+ * // Get the Auth service for a given app
+ * const otherAuth = getAuth(otherApp);
+ * ```
+ *
+ */
+export function getAuth(app?: App): Auth {
+  if (typeof app === 'undefined') {
+    app = getApp();
+  }
+
+  const firebaseApp: FirebaseApp = app as FirebaseApp;
+  return firebaseApp.getOrInitService('auth', (app) => new Auth(app));
+}
+
 export { ActionCodeSettings } from './action-code-settings-builder';
 
 export {
@@ -90,5 +124,3 @@ export {
   UserMetadata,
   UserRecord,
 } from './user-record';
-
-export { getAuth, auth } from './auth-namespace';
