+++ src/machine-learning/index.ts
@@ -69,62 +69,3 @@ export function getMachineLearning(app?: App): MachineLearning {
   const firebaseApp: FirebaseApp = app as FirebaseApp;
   return firebaseApp.getOrInitService('machineLearning', (app) => new MachineLearning(app));
 }
-
-import {
-  ListModelsResult as TListModelsResult,
-  MachineLearning as TMachineLearning,
-  Model as TModel,
-  TFLiteModel as TTFLiteModel,
-} from './machine-learning';
-import {
-  AutoMLTfliteModelOptions as TAutoMLTfliteModelOptions,
-  GcsTfliteModelOptions as TGcsTfliteModelOptions,
-  ListModelsOptions as TListModelsOptions,
-  ModelOptions as TModelOptions,
-  ModelOptionsBase as TModelOptionsBase,
-} from './machine-learning-api-client';
-
-/**
- * Gets the {@link machineLearning.MachineLearning `MachineLearning`} service for the
- * default app or a given app.
- *
- * `admin.machineLearning()` can be called with no arguments to access the
- * default app's {@link machineLearning.MachineLearning
- * `MachineLearning`} service or as `admin.machineLearning(app)` to access
- * the {@link machineLearning.MachineLearning `MachineLearning`}
- * service associated with a specific app.
- *
- * @example
- * ```javascript
- * // Get the MachineLearning service for the default app
- * var defaultMachineLearning = admin.machineLearning();
- * ```
- *
- * @example
- * ```javascript
- * // Get the MachineLearning service for a given app
- * var otherMachineLearning = admin.machineLearning(otherApp);
- * ```
- *
- * @param app Optional app whose `MachineLearning` service to
- *   return. If not provided, the default `MachineLearning` service
- *   will be returned.
- *
- * @return The default `MachineLearning` service if no app is provided or the
- *   `MachineLearning` service associated with the provided app.
- */
-export declare function machineLearning(app?: App): machineLearning.MachineLearning;
-
-/* eslint-disable @typescript-eslint/no-namespace */
-export namespace machineLearning {
-  export type ListModelsResult = TListModelsResult;
-  export type MachineLearning = TMachineLearning;
-  export type Model = TModel;
-  export type TFLiteModel = TTFLiteModel;
-
-  export type AutoMLTfliteModelOptions = TAutoMLTfliteModelOptions;
-  export type GcsTfliteModelOptions = TGcsTfliteModelOptions;
-  export type ListModelsOptions = TListModelsOptions;
-  export type ModelOptions = TModelOptions;
-  export type ModelOptionsBase = TModelOptionsBase;
-}
