+++ src/project-management/index.ts
@@ -59,60 +59,3 @@ export function getProjectManagement(app?: App): ProjectManagement {
   const firebaseApp: FirebaseApp = app as FirebaseApp;
   return firebaseApp.getOrInitService('projectManagement', (app) => new ProjectManagement(app));
 }
-
-/**
- * Gets the {@link projectManagement.ProjectManagement
- * `ProjectManagement`} service for the default app or a given app.
- *
- * `admin.projectManagement()` can be called with no arguments to access the
- * default app's {@link projectManagement.ProjectManagement
- * `ProjectManagement`} service, or as `admin.projectManagement(app)` to access
- * the {@link projectManagement.ProjectManagement `ProjectManagement`}
- * service associated with a specific app.
- *
- * @example
- * ```javascript
- * // Get the ProjectManagement service for the default app
- * var defaultProjectManagement = admin.projectManagement();
- * ```
- *
- * @example
- * ```javascript
- * // Get the ProjectManagement service for a given app
- * var otherProjectManagement = admin.projectManagement(otherApp);
- * ```
- *
- * @param app Optional app whose `ProjectManagement` service
- *     to return. If not provided, the default `ProjectManagement` service will
- *     be returned. *
- * @return The default `ProjectManagement` service if no app is provided or the
- *   `ProjectManagement` service associated with the provided app.
- */
-export declare function projectManagement(app?: App): projectManagement.ProjectManagement;
-
-import {
-  AppMetadata as TAppMetadata,
-  AppPlatform as TAppPlatform,
-} from './app-metadata';
-import { ProjectManagement as TProjectManagement } from './project-management';
-import {
-  AndroidApp as TAndroidApp,
-  AndroidAppMetadata as TAndroidAppMetadata,
-  ShaCertificate as TShaCertificate,
-} from './android-app';
-import {
-  IosApp as TIosApp,
-  IosAppMetadata as TIosAppMetadata,
-} from './ios-app';
-
-/* eslint-disable @typescript-eslint/no-namespace */
-export namespace projectManagement {
-  export type AppMetadata = TAppMetadata;
-  export type AppPlatform = TAppPlatform;
-  export type ProjectManagement = TProjectManagement;
-  export type IosApp = TIosApp;
-  export type IosAppMetadata = TIosAppMetadata;
-  export type AndroidApp = TAndroidApp;
-  export type AndroidAppMetadata = TAndroidAppMetadata;
-  export type ShaCertificate = TShaCertificate;
-}
