+++ src/remote-config/index.ts
@@ -71,67 +71,3 @@ export function getRemoteConfig(app?: App): RemoteConfig {
   const firebaseApp: FirebaseApp = app as FirebaseApp;
   return firebaseApp.getOrInitService('remoteConfig', (app) => new RemoteConfig(app));
 }
-
-import {
-  ExplicitParameterValue as TExplicitParameterValue,
-  InAppDefaultValue as TInAppDefaultValue,
-  ListVersionsOptions as TListVersionsOptions,
-  ListVersionsResult as TListVersionsResult,
-  RemoteConfigCondition as TRemoteConfigCondition,
-  RemoteConfigParameter as TRemoteConfigParameter,
-  RemoteConfigParameterGroup as TRemoteConfigParameterGroup,
-  RemoteConfigParameterValue as TRemoteConfigParameterValue,
-  RemoteConfigTemplate as TRemoteConfigTemplate,
-  RemoteConfigUser as TRemoteConfigUser,
-  TagColor as TTagColor,
-  Version as TVersion,
-} from './remote-config-api';
-import { RemoteConfig as TRemoteConfig } from './remote-config';
-
-/**
- * Gets the {@link remoteConfig.RemoteConfig `RemoteConfig`} service for the
- * default app or a given app.
- *
- * `admin.remoteConfig()` can be called with no arguments to access the default
- * app's {@link remoteConfig.RemoteConfig `RemoteConfig`} service or as
- * `admin.remoteConfig(app)` to access the
- * {@link remoteConfig.RemoteConfig `RemoteConfig`} service associated with a
- * specific app.
- *
- * @example
- * ```javascript
- * // Get the `RemoteConfig` service for the default app
- * var defaultRemoteConfig = admin.remoteConfig();
- * ```
- *
- * @example
- * ```javascript
- * // Get the `RemoteConfig` service for a given app
- * var otherRemoteConfig = admin.remoteConfig(otherApp);
- * ```
- *
- * @param app Optional app for which to return the `RemoteConfig` service.
- *   If not provided, the default `RemoteConfig` service is returned.
- *
- * @return The default `RemoteConfig` service if no
- *   app is provided, or the `RemoteConfig` service associated with the provided
- *   app.
- */
-export declare function remoteConfig(app?: App): remoteConfig.RemoteConfig;
-
-/* eslint-disable @typescript-eslint/no-namespace */
-export namespace remoteConfig {
-  export type ExplicitParameterValue = TExplicitParameterValue;
-  export type InAppDefaultValue = TInAppDefaultValue;
-  export type ListVersionsOptions = TListVersionsOptions;
-  export type ListVersionsResult = TListVersionsResult;
-  export type RemoteConfig = TRemoteConfig;
-  export type RemoteConfigCondition = TRemoteConfigCondition;
-  export type RemoteConfigParameter = TRemoteConfigParameter;
-  export type RemoteConfigParameterGroup = TRemoteConfigParameterGroup;
-  export type RemoteConfigParameterValue = TRemoteConfigParameterValue;
-  export type RemoteConfigTemplate = TRemoteConfigTemplate;
-  export type RemoteConfigUser = TRemoteConfigUser;
-  export type TagColor = TTagColor;
-  export type Version = TVersion;
-}
