+++ src/security-rules/index.ts
@@ -26,14 +26,6 @@ export {
   SecurityRules,
 } from './security-rules';
 
-import {
-  RulesFile as TRulesFile,
-  Ruleset as TRuleset,
-  RulesetMetadata as TRulesetMetadata,
-  RulesetMetadataList as TRulesetMetadataList,
-  SecurityRules as TSecurityRules,
-} from './security-rules';
-
 /**
  * Gets the {@link securityRules.SecurityRules
  * `SecurityRules`} service for the default app or a given app.
@@ -69,43 +61,4 @@ export function getSecurityRules(app?: App): SecurityRules {
 
   const firebaseApp: FirebaseApp = app as FirebaseApp;
   return firebaseApp.getOrInitService('securityRules', (app) => new SecurityRules(app));
-}
-
-/**
- * Gets the {@link securityRules.SecurityRules
- * `SecurityRules`} service for the default app or a given app.
- *
- * `admin.securityRules()` can be called with no arguments to access the
- * default app's {@link securityRules.SecurityRules
- * `SecurityRules`} service, or as `admin.securityRules(app)` to access
- * the {@link securityRules.SecurityRules `SecurityRules`}
- * service associated with a specific app.
- *
- * @example
- * ```javascript
- * // Get the SecurityRules service for the default app
- * var defaultSecurityRules = admin.securityRules();
- * ```
- *
- * @example
- *  ```javascript
- * // Get the SecurityRules service for a given app
- * var otherSecurityRules = admin.securityRules(otherApp);
- * ```
- *
- * @param app Optional app to return the `SecurityRules` service
- *     for. If not provided, the default `SecurityRules` service
- *     is returned.
- * @return The default `SecurityRules` service if no app is provided, or the
- *   `SecurityRules` service associated with the provided app.
- */
-export declare function securityRules(app?: App): securityRules.SecurityRules;
-
-/* eslint-disable @typescript-eslint/no-namespace */
-export namespace securityRules {
-  export type RulesFile = TRulesFile;
-  export type Ruleset = TRuleset;
-  export type RulesetMetadata = TRulesetMetadata;
-  export type RulesetMetadataList = TRulesetMetadataList;
-  export type SecurityRules = TSecurityRules;
 }
