+++ src/security-rules/security-rules-namespace.ts
@@ -1,0 +1,63 @@
+/*!
+ * Copyright 2021 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { App } from '../app';
+import {
+  RulesFile as TRulesFile,
+  Ruleset as TRuleset,
+  RulesetMetadata as TRulesetMetadata,
+  RulesetMetadataList as TRulesetMetadataList,
+  SecurityRules as TSecurityRules,
+} from './security-rules';
+
+/**
+ * Gets the {@link securityRules.SecurityRules
+  * `SecurityRules`} service for the default app or a given app.
+  *
+  * `admin.securityRules()` can be called with no arguments to access the
+  * default app's {@link securityRules.SecurityRules
+  * `SecurityRules`} service, or as `admin.securityRules(app)` to access
+  * the {@link securityRules.SecurityRules `SecurityRules`}
+  * service associated with a specific app.
+  *
+  * @example
+  * ```javascript
+  * // Get the SecurityRules service for the default app
+  * var defaultSecurityRules = admin.securityRules();
+  * ```
+  *
+  * @example
+  *  ```javascript
+  * // Get the SecurityRules service for a given app
+  * var otherSecurityRules = admin.securityRules(otherApp);
+  * ```
+  *
+  * @param app Optional app to return the `SecurityRules` service
+  *     for. If not provided, the default `SecurityRules` service
+  *     is returned.
+  * @return The default `SecurityRules` service if no app is provided, or the
+  *   `SecurityRules` service associated with the provided app.
+  */
+export declare function securityRules(app?: App): securityRules.SecurityRules;
+
+/* eslint-disable @typescript-eslint/no-namespace */
+export namespace securityRules {
+  export type RulesFile = TRulesFile;
+  export type Ruleset = TRuleset;
+  export type RulesetMetadata = TRulesetMetadata;
+  export type RulesetMetadataList = TRulesetMetadataList;
+  export type SecurityRules = TSecurityRules;
+}
